% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.listPtrv.r
\name{analysis.ptrvList}
\alias{analysis.ptrvList}
\title{analysis.listPtrv}
\usage{
\method{analysis}{ptrvList}(
  x,
  type,
  ionToRemove = NULL,
  normalizeByEval = TRUE,
  log = FALSE,
  outputFormat = "",
  outputFile = "",
  title = "",
  runBy = "",
  selection = "",
  ...
)
}
\arguments{
\item{x}{Object of class "tds" or "fc" or "profile".}

\item{type}{Choose among: "CVA of abundances","ANOVA of intensities","PCA of abundances","Intensity Heatmap"}

\item{ionToRemove}{vector of character containing ions to remove from the analysis}

\item{normalizeByEval}{Boolean. By default TRUE}

\item{log}{boolean indicated if the intensities should be logged (TRUE) or not (FALSE)}

\item{outputFormat}{Format of output, "" (no output), "html", "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp, "svg" or "wmf" (windows only)}

\item{outputFile}{Name of the file (if outputFormat !="").}

\item{title}{Title of the output.}

\item{runBy}{Split x in several datasets corresponding to the combinations of each factor of runBy, then run a separate analysis for each dataset.}

\item{selection}{Run the analysis for a subset of x. Use syntax of sqldf.}

\item{...}{Further analysis parameters}
}
\description{
analysis.listPtrv
}
