% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Generic function}
\usage{
analysis(
  x,
  outputFormat = "",
  outputFile = "output",
  title = "analysis",
  runBy = "",
  selection = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class "tds" or "fc" or "profile".}

\item{outputFormat}{Format of output, "" (no output), "html", "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp, "svg" or "wmf" (windows only)}

\item{outputFile}{Name of the file (if outputFormat !="").}

\item{title}{Title of the output.}

\item{runBy}{Split x in several datasets corresponding to the combinations of each factor of runBy, then run a separate analysis for each dataset.}

\item{selection}{Run the analysis for a subset of x. Use syntax of sqldf.}

\item{...}{Further parameters}
}
\value{
A list, with elements depending on type (see details).
}
\description{
Generic function.
}
