% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaTable.r
\name{anovaTable}
\alias{anovaTable}
\title{Run a batch of ANOVAs and store the results in a table.}
\usage{
anovaTable(
  df,
  model,
  alpha = 0.05,
  unit = "descriptor",
  columns = c("G_Mean", "F_Product", "P_Product", "F_Product_Significance", "SE_Product",
    "R2", "P_ShapiroWilks", "P_Levene_Product", "Mean_Product", "Group_Product")
)
}
\arguments{
\item{df}{A dataframe with columns subject, product, "unit", score.}

\item{model}{A lmer model of ANOVA.}

\item{alpha}{Risk for the post-hoc multiple comparison test.}

\item{unit}{The name of the column of the dataframe that will be used as a loop.}

\item{columns}{Elements to add in the table.}
}
\value{
A dataframe.
}
\description{
Run a batch of ANOVAs and store the results in a table.
}
\examples{
{
#data(duration)
#anovaTable(duration,model='score~product')
}
}
