% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectStartsInFile.r
\name{detectStartsInFile}
\alias{detectStartsInFile}
\title{Detects several starts in one file (according to a vector of theoretical stimulation times)}
\usage{
detectStartsInFile(
  ptrFile_i,
  ion,
  startPeaks,
  endPeaks = NULL,
  methods = c("higherThanNoise", "startPeakProportion", "higherDerivative"),
  nPoints = 31,
  multiplyNoiseBy = 1.5,
  statOfNoise = "bl",
  smooth = TRUE,
  proportionOfMax = 0.1,
  peakChoice = "maxIntensity",
  order = 2,
  detectionThreshold = 15,
  firstAmongHighThreshold = 50,
  manual = NULL
)
}
\arguments{
\item{ptrFile_i}{data frame exported from ptrViewer}

\item{ion}{name of the column corresponding to the ion to analyse}

\item{startPeaks}{a vector containing the starts for all files}

\item{endPeaks}{if NULL, the end of the peak period is given by the next start in startPeaks}

\item{methods}{Choose among c("higherThanNoise","startPeakProportion","higherDerivative")}

\item{nPoints}{number of points required in the smoothing function}

\item{multiplyNoiseBy}{required if method='higherThanNoise' is chosen}

\item{statOfNoise}{Required if method='higherThanNoise'. Statistic to be multiplied by multiplyByNoise ('max' or 'avg'). Can also be 'bl' if corresponding to the estimation of noise during the whole evaluation or 'blper' if}

\item{smooth}{If TRUE, the data is previously smoothed with the parameter halfWindowsSize}

\item{proportionOfMax}{Number between 0 and 1. If method startPeakProportion is chosen and length(starts)==1, the starting time will be the first time that the intensity reaches this proportion of the maximale intensity of the peak. If length(starts) > 1, the starting time will be the minimum value obtained with this method for all starts.}

\item{peakChoice}{"maxIntensity", "first" or "firstAmongHigh"}

\item{order}{order of the derivative}

\item{detectionThreshold}{intensity value to be reached to be considered as a peak (default to NULL corresponds to 0)}

\item{firstAmongHighThreshold}{percentage of the maximum value to be reached in order to be considered as a peak when peak choice is firstAmongHigh}

\item{manual}{manual results obtained (as a vector of the same length as startPeaks)}

\item{rit}{}
}
\value{
a list containing dfres (dataframe with the obtained times), gg (a graphic representing the obtained intensities and starting time points, diagnosis)
}
\description{
Detects several starts in one file (according to a vector of theoretical stimulation times)
}
