% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.r, R/diagnosisNamefile.r
\name{diagnosis}
\alias{diagnosis}
\title{diagnosis
Return whether a list of files (in a specific repository) has the same number and the same name of columns (required to run ptrvList functions)}
\usage{
diagnosis(listFiles, sep = sep)

diagnosis(listFiles, sep = sep)
}
\arguments{
\item{listFiles}{the list of files to be read}

\item{sep}{separator of the files}
}
\value{
a list containing ncol (the number of column in each file) and colnames (the colnames in each file). They should be identical.
}
\description{
diagnosis
Return whether a list of files (in a specific repository) has the same number and the same name of columns (required to run ptrvList functions)
}
