% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByEval.r
\name{normalizeByEval}
\alias{normalizeByEval}
\title{@param intens data frame containing the column "file","product","subject","rep","ion","intensity" (optionally "file" and "time")
 @param ionToUse if NULL, all the ions are use to normalize, if a vector of names of ions, it is used to normalize
 @export}
\usage{
normalizeByEval(intens, ionToUse = NULL)
}
\description{
@param intens data frame containing the column "file","product","subject","rep","ion","intensity" (optionally "file" and "time")
 @param ionToUse if NULL, all the ions are use to normalize, if a vector of names of ions, it is used to normalize
 @export
}
