% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickingPeaks.r
\name{pickingPeaks}
\alias{pickingPeaks}
\title{pickingPeaks}
\usage{
pickingPeaks(
  df,
  col = "mz",
  signalPercentage = 10,
  nPoints = 0,
  refineMz = "descendPeak",
  method = "MAD"
)
}
\arguments{
\item{df}{a dataframe containing two columns: "mz" and "intensity" corresponding respectively to abscissa and ordinates.}

\item{col}{name of the column containing the abscissa ("mz" by default)}

\item{signalPercentage}{of the centroid's intensity are used to calculate the refined m/z. By default the descend is stopped when the first signal that is equal or larger than the last observed one is encountered.}
}
\description{
pickingPeaks
}
