% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lcmsHeatmap.R
\name{plot.hm}
\alias{plot.hm}
\title{plot_lcmsHeatmap}
\usage{
\method{plot}{hm}(x, main = "pheatmap", log = FALSE, scale_limits = NULL, ...)
}
\arguments{
\item{x}{result of lcmsHeatmap}

\item{main}{title of the graph}

\item{log}{if TRUE, the data is logtransformed before plotting}

\item{scale_limits}{a numeric vector of size two for the limits of the color scale}

\item{...}{Further parameters in plots such as xlim,xlab,ylim,ylab}
}
\description{
plot_lcmsHeatmap
}
