% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ibm.R
\name{plot.ibm}
\alias{plot.ibm}
\title{plot.ibm}
\usage{
\method{plot}{ibm}(
  x,
  subset = NULL,
  main = "Mass spectra",
  xlab = "Mass",
  ylab = "Intensity",
  groupOption = "no",
  xlim = NULL,
  ylim = NULL,
  type = "l",
  ...
)
}
\arguments{
\item{x}{an ibm object (from intensityByMass functions)}

\item{subset}{vector containing some names present in the "name" column}

\item{main}{title of the graph}

\item{xlab}{character indicating the label on the x axis}

\item{ylab}{character indicating the label on the y axis}

\item{groupOption}{option for grouping the data "name" or "no"}

\item{xlim}{vector of two numbers representing the limits of x axis.}

\item{ylim}{vector of two numbers representing the limits of y axis.}

\item{type}{"l" for line, "bar" for barplot}

\item{...}{Further parameters in plots such as xlim,xlab,ylim,ylab}
}
\description{
plot.ibm
}
\examples{
#data(lcms)
#res=lcmsIntensityByMass(lcms)
#plot(res)
}
