% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ibt.R
\name{plot.ibt}
\alias{plot.ibt}
\title{plot.ibt}
\usage{
\method{plot}{ibt}(
  x,
  main = "Chromatogram",
  xlab = "time",
  ylab = "intensity",
  ylim = NULL,
  xlim = NULL,
  selec = NULL,
  groupOption = "no",
  ...
)
}
\arguments{
\item{x}{abscissa of the spectra}

\item{main}{title of the graph}

\item{xlab}{character indicating the label on the x axis}

\item{ylab}{character indicating the label on the y axis}

\item{ylim}{vector of two numbers representing the limits of y axis.}

\item{xlim}{vector of two numbers representing the limits of x axis.}

\item{selec}{vector containing the names of the spectra to be displayed}

\item{groupOption}{"no", "name" or file.}

\item{...}{Further parameters in plots such as xlim,xlab,ylim,ylab}
}
\description{
plot.ibt
}
