% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fullspectra.r
\name{plot_fullspectra}
\alias{plot_fullspectra}
\title{This function allows to draw spectra according to a group and potential smoothing}
\usage{
plot_fullspectra(
  x,
  y,
  group = NULL,
  spar = 0,
  integrationTable = NULL,
  xlim = NULL,
  ylim = NULL,
  threshold = 0,
  main = "Spectra",
  xlab = "Mass",
  ylab = "Intensity",
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{abscissa of the spectra}

\item{y}{ordinate of the spectra}

\item{group}{group for each spectra}

\item{spar}{if 0, no smoothing. Ifelse the smoothing parameter is between 0 and 1 and indicates a smoothing.}

\item{integrationTable}{NULL}

\item{xlim}{vector of two numbers representing the limits of x axis.}

\item{ylim}{vector of two numbers representing the limits of y axis.}

\item{threshold}{NULL}

\item{main}{title of the graph}

\item{xlab}{character indicating the label on the x axis}

\item{ylab}{character indicating the label on the y axis}

\item{legend}{if TRUE the legend is displayed}

\item{...}{Further parameters in plots such as xlim,xlab,ylim,ylab}
}
\description{
This function allows to draw spectra according to a group and potential smoothing
}
