% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrIntensityByMass.r
\name{ptrIntensityByMass}
\alias{ptrIntensityByMass}
\title{ptrIntensityByMass}
\usage{
ptrIntensityByMass(
  ptr,
  type = "exactMass",
  rt = NULL,
  integrationTable = NULL,
  breaks = NULL,
  by = 0.1
)
}
\arguments{
\item{ptr}{ptr object stemming from ptrRead}

\item{type}{file .h5 to read}

\item{rt}{vector of two numbers (min time and max time). If NULL (default), no time limits.}

\item{integrationTable}{integrationTable containing columns: "mz","inf" and "sup"}

\item{breaks}{breaks to be used}

\item{by}{step of time for integration of the chromatogram. Default 0.1}
}
\description{
ptrIntensityByMass
}
\examples{
data(ptr)
#ptrIntensityByMass(ptr)
}
