% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrIntensityByTime.r
\name{ptrIntensityByTime}
\alias{ptrIntensityByTime}
\title{ptrIntensityByTime}
\usage{
ptrIntensityByTime(ptr, rt = NULL, mz = NULL, integrationTable = NULL)
}
\arguments{
\item{ptr}{a ptr object (from ptrRead function)}

\item{rt}{A numeric(2) or two-column matrix defining the lower and upper boundary for the retention time range/window(s) for the chromatogram(s). If a matrix is provided, a chromatogram is extracted for each row. If not specified, a chromatogram representing the full retention time range is extracted.}

\item{mz}{A numeric(2) or two-column matrix defining the mass-to-charge (mz) range(s) for the chromatogram(s). For each spectrum/retention time, all intensity values within this mz range are aggregated to result in the intensity value for the spectrum/retention time. If not specified, the full mz range is considered.}

\item{integrationTable}{If not NULL, gives the intensity by time according to an integrationTable}
}
\description{
ptrIntensityByTime

ptrvIntensity
}
