% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrRead.r
\name{ptrRead}
\alias{ptrRead}
\title{Read .h5 file from PTR and return a list containing useful results}
\usage{
ptrRead(
  file,
  wd = getwd(),
  mz = NULL,
  threshold = 0,
  sumSpectraOnly = FALSE,
  integrationTable = NULL
)
}
\arguments{
\item{file}{file .h5 to read}

\item{wd}{path for the working directory}

\item{mz}{If not NULL, vector of length 2 representing a relevant interval of mz to be studied}

\item{threshold}{threshold}

\item{sumSpectraOnly}{load only sum spectra}

\item{integrationTable}{integrationTable}
}
\description{
Read .h5 file from PTR and return a list containing useful results
}
\examples{
#DONOTRUN
#file <- dir(system.file(package = "chemosensR", dir = "extdata"),full.name=TRUE,pattern="h5$")
#ptr=ptrRead(file)
}
