% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrReadListIBM.r
\name{ptrReadListIBM}
\alias{ptrReadListIBM}
\title{ptrReadListIBM
This function aims to read a list of files of PTR type and to return an ibm object containing a dataframe with "file", "mz","intensity","file" columns}
\usage{
ptrReadListIBM(
  listFiles,
  sumSpectraOnly = TRUE,
  normalization = "none",
  rt = NULL,
  integrationTable = NULL,
  breaks = NULL,
  by = 0.1,
  type = "sumSpectrum"
)
}
\arguments{
\item{listFiles}{vector of characters representing the names of the files to be loaded}

\item{sumSpectraOnly}{TRUE by default. Faster if no time is required in the analysis.}

\item{normalization}{"none" # TODO}

\item{rt}{vector of two numbers (min time and max time). If NULL (default), no time limits.}

\item{integrationTable}{integrationTable containing columns: "mz","inf" and "sup"}

\item{breaks}{breaks to be used}

\item{by}{step of time for integration of the chromatogram. Default 0.1}

\item{type}{file .h5 to read}
}
\value{
a ibm object
}
\description{
ptrReadListIBM
This function aims to read a list of files of PTR type and to return an ibm object containing a dataframe with "file", "mz","intensity","file" columns
}
\examples{
# files <- dir(system.file(package = "chemosensR", dir = "extdata"),full.name=TRUE,pattern="h5$")
# ptr_ibm=ptrReadListIBM(files)
}
