% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrReadListIBT.r
\name{ptrReadListIBT}
\alias{ptrReadListIBT}
\title{This function aims to read a list of files of PTR type and to return an ibm object containing a dataframe with "name", "mz","intensity","file" columns}
\usage{
ptrReadListIBT(
  listFiles,
  sumSpectraOnly = TRUE,
  mz = NULL,
  rt = NULL,
  integrationTable = NULL
)
}
\arguments{
\item{listFiles}{vector of characters representing the names of the files to be loaded}

\item{sumSpectraOnly}{TRUE by default. Faster if no time is required in the analysis.}

\item{mz}{A numeric(2) or two-column matrix defining the mass-to-charge (mz) range(s) for the chromatogram(s). For each spectrum/retention time, all intensity values within this mz range are aggregated to result in the intensity value for the spectrum/retention time. If not specified, the full mz range is considered.}

\item{rt}{A numeric(2) or two-column matrix defining the lower and upper boundary for the retention time range/window(s) for the chromatogram(s). If a matrix is provided, a chromatogram is extracted for each row. If not specified, a chromatogram representing the full retention time range is extracted.}

\item{integrationTable}{If not NULL, gives the intensity by time according to an integrationTable}
}
\value{
a ibt object
}
\description{
This function aims to read a list of files of PTR type and to return an ibm object containing a dataframe with "name", "mz","intensity","file" columns
}
