% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrTrajectoryPca.r
\name{ptrTrajectoryPca}
\alias{ptrTrajectoryPca}
\title{Trajectory PCA}
\usage{
ptrTrajectoryPca(
  tds,
  axes = list(c(1, 2)),
  title = "Trajectory PCA",
  periods = 10,
  individual = "arrows",
  fontSizeCex = 1,
  smooth = FALSE,
  expandBiplot = NULL,
  parameters = NULL
)
}
\arguments{
\item{tds}{a tds object \link[chemosensR]{tdsRead}}

\item{axes}{list of pairs of axes (for example list(c(1,2),c(1,3),c(2,3)). Default to list(c(1,2))}

\item{title}{title of the graph}

\item{periods}{number of periods required in the cutting in time-periods}

\item{individual}{representation of individuals Default to "arrows"}

\item{fontSizeCex}{size of graphical textual elements (integers)}

\item{smooth}{TRUE if the data should be smoothed, FALSE ifelse. Default to FALSE}
}
\value{
A list containing the coordinates of PCA, then output containing the ggplot object required for the plot
}
\description{
Computes a PCA of trajectories. Time is decomposed in time periods. The evaluation*periods are then used as individuals to produce a PCA.  These individuals are plotted then linked by a segments for successive periods.
}
