% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvComparableData_one.r
\name{ptrvComparableData}
\alias{ptrvComparableData}
\title{ptrvComparableData}
\usage{
ptrvComparableData(df, miseEnBouche = 0, stopTime = Inf, timeColumn = "time")
}
\arguments{
\item{df}{data.frame with time column to be translated}

\item{miseEnBouche}{Number corresponding to the starting time of tasting}

\item{stopTime}{End of the tasting time}

\item{timeColumn}{Name of the column time by default (="time")}
}
\description{
ptrvComparableData
}
