% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvCreateMetafile.r
\name{ptrvCreateMetaFile}
\alias{ptrvCreateMetaFile}
\title{ptrvCreateMetaFile}
\usage{
ptrvCreateMetaFile(
  wd,
  subject = NULL,
  product = NULL,
  replicate = NULL,
  sep = ";"
)
}
\arguments{
\item{wd}{Working directory to use for the creation of the metafile}

\item{subject}{vector of length 2 containing the position of character corresponding to the subject in the file}

\item{product}{vector of length 2 containing the position of character corresponding to the subject in the file}

\item{replicate}{vector of length 2 containing the position of character corresponding to the subject in the file}

\item{sep}{vector of separators for reading the files =";",}
}
\description{
ptrvCreateMetaFile
}
