% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvDetectStart.r
\name{ptrvDetectStart}
\alias{ptrvDetectStart}
\title{ptrvDetectStart}
\usage{
ptrvDetectStart(
  res,
  starts,
  method = "startPeakProportion",
  proportionOfMax = 0.1,
  multiplyNoiseBy = 3,
  peakChoice = "maxIntensity",
  statOfNoise = "max",
  noisePeriod = c(0, 20),
  timeChoice = "interpolation",
  startPeriod = c(20, 50),
  smooth = FALSE,
  halfWindowSize = 3,
  nPoints = 7,
  order = 1,
  firstAmongHighThreshold = 50
)
}
\arguments{
\item{res}{result of ptrvIntensityByTime$res (data.frame with time, intensity and ion as columns)}

\item{starts}{name of the ions to be used to detect the starting point}

\item{method}{"startPeakProportion" "higherThanNoise" "tangential" TODO}

\item{proportionOfMax}{Number between 0 and 1. If method startPeakProportion is chosen and length(starts)==1, the starting time will be the first time that the intensity reaches this proportion of the maximale intensity of the peak. If length(starts) > 1, the starting time will be the minimum value obtained with this method for all starts.}

\item{multiplyNoiseBy}{Number higher than 1 (nb). If method "higherThanNoise" is chosen, the maximal (or averaged according to noiseStat) value obtained during the noise period is multiplied by this number to give the intensity to reach for returning the starting period.}

\item{statOfNoise}{"avg" or "max". If method "higherThanNoise" is chosen, the statistic obtained during the noise period to be multiplied by multiplyNoiseBy for returning the starting period.}

\item{noisePeriod}{Vector of two numbers. If method "higherThanNoise" is chosen and length(starts)==1, the starting time will be the first time that the intensity reaches 3*the maximal value obtained during the noise period. If length(starts) > 1, the starting time will be the minimum value obtained with this method for all starts.}

\item{timeChoice}{"interpolation", "lower","upper". The time obtained for reaching a given value can be a not observed time (but in a interval of two observed times). "interpolation" returns the not observed time, "lower" returns the lower limit of this interval, "upper" returns the upper limit of this interval}

\item{startPeriod}{Vector of two numbers. The starting time obtained has to be in this interval.}

\item{smooth}{If TRUE, the data is previously smoothed with the parameter halfWindowsSize}

\item{halfWindowSize, }{half window size used for smoothing}
}
\description{
ptrvDetectStart
}
\examples{
data(ptrv)
res_intensity=ptrvIntensityByTime(ptrv,referenceBreath="m69.06989..isoprene...Conc.",
correction = "cycle",removeNoise=FALSE,breathRatio = FALSE)
ion="m115.11139..C7H15O....Conc."
ds=ptrvDetectStart(res=res_intensity$res,starts=ion, method="startPeakProportion")
ds=ptrvDetectStart(res=res_intensity$res,starts=ion, method="higherThanNoise",startPeriod=c(20,150))
ds=ptrvDetectStart(res=res_intensity$res,starts=ion, method="higherThanNoise",statOfNoise="avg",
multiplyNoiseBy=1.6,timeChoice="lower")
ds=ptrvDetectStart(res=res_intensity$res,starts=ion, method="higherThanNoise",statOfNoise="avg",
multiplyNoiseBy=1.6,smooth=TRUE,startPeriod=c(20,150))
ds=ptrvDetectStart(res=res_intensity$res,starts=ion, method="startPeakProportion",
proportionOfMax=0.1,smooth=TRUE,startPeriod=c(20,150))
}
