% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvIntensity.r
\name{ptrvIntensity}
\alias{ptrvIntensity}
\title{ptrvIntensity}
\usage{
ptrvIntensity(
  dataset,
  timePeriod = NULL,
  negativeAsNull = TRUE,
  propPeak = FALSE,
  proportion = 0.75,
  timing = "last"
)
}
\arguments{
\item{dataset}{is a result of ptrvIntensityByTime$res: a dataframe with colnames as 'time','duration','ion' and  'intensity'}

\item{timePeriod}{vector with two numbers indicating an interval to calculate the intensities}

\item{negativeAsNull}{If TRUE, all negative value are replaced by 0}
}
\description{
ptrvIntensity
}
