% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListAnova.r
\name{ptrvListAnova}
\alias{ptrvListAnova}
\title{ptrvListAnova
returns the two way mixed ANOVA product*subject (with random subject effect) for each ion}
\usage{
ptrvListAnova(
  df,
  ionToRemove = NULL,
  log = FALSE,
  normalizeByEval = FALSE,
  model = "intensity~product+(1|subject)+(1|product:subject)",
  alpha = 0.05,
  columns = c("G_Mean", "F_Product", "P_Product", "F_Product_Significance",
    "SE_Product", "R2", "P_ShapiroWilks", "P_Levene_Product", "Mean_Product",
    "Group_Product", "P_Subject")
)
}
\arguments{
\item{df}{data frame with "ion","intensity","product","subject" columns,}

\item{ionToRemove}{vector of ion to remove from the analysis}

\item{log}{if TRUE the data is logged}

\item{normalizeByEval}{if TRUE, each evaluation (row) of the matrix is divided by the sum of all the variables for this evaluation (row sum)}

\item{model}{"intensity~product+(1|subject)+(1|product:subject)" or other model containing these factors}

\item{alpha}{limit of significance}

\item{columns}{choose between c("G_Mean","F_Product","P_Product","F_Product_Significance","SE_Product","R2","P_ShapiroWilks","P_Levene_Product","Mean_Product","Group_Product","P_Subject") )}
}
\description{
ptrvListAnova
returns the two way mixed ANOVA product*subject (with random subject effect) for each ion
}
