% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListCVA.r
\name{ptrvListCVA}
\alias{ptrvListCVA}
\title{ptrvListCVA}
\usage{
ptrvListCVA(
  df,
  ionToRemove = NULL,
  option = "OneWayANOVA",
  normalizeByEval = FALSE,
  log = FALSE,
  axes = c(1, 2)
)
}
\arguments{
\item{df}{data frame with "ion","intensity","product","subject" columns,}

\item{ionToRemove}{vector of character containing ions to remove from the analysis}

\item{option}{"OneWayANOVA" or "TwoWayANOVA"}

\item{normalizeByEval}{Boolean. By default TRUE}

\item{log}{boolean indicated if the intensities should be logged (TRUE) or not (FALSE)}

\item{axes}{a list of couples of integers corresponding to the maps to be obtained}
}
\description{
ptrvListCVA
}
