% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListHeatmap.r
\name{ptrvListHeatmap}
\alias{ptrvListHeatmap}
\title{ptrvListHeatmap
returns heatmap with rows as individuals and columns as ions}
\usage{
ptrvListHeatmap(
  df,
  log = FALSE,
  normalizeByEval = FALSE,
  ionToRemove = NULL,
  formula = as.formula(product + subject + rep ~ ion),
  fun.aggregate = "mean",
  breaks = NULL,
  clusterRows = F,
  clusterCols = F,
  showRownames = F,
  showColnames = T,
  annotationRow = "product",
  scale = "column",
  cex = 0.8,
  transpose = FALSE
)
}
\arguments{
\item{df}{data frame with "ion","intensity","product","subject" columns}

\item{log}{boolean indicated if the intensities should be logged (TRUE) or not (FALSE)}

\item{normalizeByEval}{Boolean. By default TRUE}

\item{ionToRemove}{vector of character containing ions to remove from the analysis}

\item{formula}{as.formula(product+subject+rep~ion) formula to be taken into account}

\item{fun.aggregate}{"mean", "max" or "sum" for regrouping according to the frmula}

\item{breaks}{a sequence of numbers that covers the range of values in mat and is one element longer than color vector. Used for mapping values to colors. Useful, if needed to map certain values to certain colors, to certain values. If value is NA then the breaks are calculated automatically. When breaks do not cover the range of values, then any value larger than max(breaks) will have the largest color and any value lower than min(breaks) will get the lowest color.}

\item{clusterRows}{Boolean indicating whether the rows should be clustered}

\item{clusterCols}{Boolean indicating whether the columns should be clustered}

\item{showRownames}{boolean specifying if row names are be shown.}

\item{showColnames}{boolean specifying if column names are be shown.}

\item{annotationRow}{"product" or "subject" (annotation shown on the left side fo the heatmap)}

\item{scale}{"none","column" or "row". Character indicating if the values should be centered and scaled in either the row direction or the column direction, or none.}

\item{cex}{integer. Size of the text.}

\item{transpose}{boolean specifying if the matrix is previously transposed (TRUE) or not (FALSE)}
}
\description{
ptrvListHeatmap
returns heatmap with rows as individuals and columns as ions
}
