% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListIntensity.r
\name{ptrvListIntensity}
\alias{ptrvListIntensity}
\title{ptrvListIntensity}
\usage{
ptrvListIntensity(
  listFiles,
  metaData = NULL,
  format = "wide",
  correction = "cycle",
  stat = "area"
)
}
\arguments{
\item{listFiles}{vector containing the file names of the files to be read}

\item{metaData}{metaData required #TODO}

\item{format}{"wide" by default}

\item{correction}{"cycle" by default.}

\item{stat}{"area" by default but can also be tmax or max}
}
\description{
ptrvListIntensity
}
