% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListIntensityByTime.r
\name{ptrvListIntensityByTime}
\alias{ptrvListIntensityByTime}
\title{ptrvListIntensityByTime
A function automizing all pretreatements for}
\usage{
ptrvListIntensityByTime(
  listFiles,
  removeBlankTime = FALSE,
  ions = NULL,
  dec_vec = rep(".", length(listFiles)),
  sep = "\\t",
  metaData = NULL,
  correction = "cycle",
  timeBlank = c(0, 30),
  halfWindowSize = 5,
  maxPeaks = NULL,
  method = "MAD",
  total = FALSE,
  breathRatio = FALSE,
  stat = "area",
  minimalDuration = 2,
  smoothMethod = "MovingAverage"
)
}
\arguments{
\item{listFiles}{list of the file names to be imported}

\item{removeBlankTime}{boolean. If TRUE, the time of blank time is substracted from the time column}

\item{ions}{vector of ion to be analyzed}

\item{dec_vec}{Vector with the same length as listFiles containing the decimal separator used in the list of files}

\item{sep}{separator used in the files to be read}

\item{metaData}{A data.frame with colnames as "file","product","subject","rep","breathing","into","swallow","end")}

\item{correction}{"none" or "cycle". See Details.}

\item{timeBlank}{vector of two numbers. Indicates the interval to take as a blank period.}

\item{halfWindowSize}{parameter used to determine the smooth for the peak-picking used in break detection}

\item{maxPeaks}{A number. Below this value, a peak is not considered as a peak}

\item{method}{method used for peak picking ("MAD" by default)}

\item{total}{if TRUE, the total intensity is calculated and added as a supplementary column}

\item{breathRatio}{if TRUE, the ratio between the intensity of ion and the breath is calculated}

\item{stat}{Statistic to be extracted for each ion in each file. Among c("sum","moy","sd","n","min","max","tmax","area")}
}
\value{
a list containing : total intensity: AUC (or tmax, etc) in a long format, time: data containing the corrected timepoints, listRes: AUC in a wide format
}
\description{
ptrvListIntensityByTime
A function automizing all pretreatements for
}
