% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListPca.r
\name{ptrvListPCA}
\alias{ptrvListPCA}
\title{Runs PCA on a data frame stemming from ptrvReadList()}
\usage{
ptrvListPCA(
  df,
  ionToRemove = NULL,
  dataType = "productMeans",
  normalizeByEval = FALSE,
  log = FALSE,
  axes = list(c(1, 2)),
  representation = "TwoMaps",
  option = "Correlation",
  variableName = "ion",
  scoreName = "intensity",
  expandBiplot = NULL
)
}
\arguments{
\item{df}{data frame with "ion","intensity","product","subject" columns,}

\item{ionToRemove}{vector of character containing ions to remove from the analysis}

\item{dataType}{"raw" (if PCA should be computed on the whole dataset) or "productMeans" if PCA should be computed on the product mean table}

\item{normalizeByEval}{Boolean. By default TRUE.}

\item{log}{boolean indicated if the intensities should be logged (TRUE) or not (FALSE)}

\item{axes}{list of couples of integers corresponding to the maps to plot}

\item{representation}{"twoMaps" or "distanceBiplot"}

\item{option}{"Correlation" or "Covariance"}

\item{variableName}{column containing the variable names. Default to 'ion'}

\item{scoreName}{column name of the score. Default to 'intensity'}

\item{expandBiplot}{a number corresponding to the expansion of the variables compared to the individuals in the biplot.}
}
\description{
Runs PCA on a data frame stemming from ptrvReadList()
}
