% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvQCAverages.r
\name{ptrvQCAverages}
\alias{ptrvQCAverages}
\title{ptrvQCAverages}
\usage{
ptrvQCAverages(df, ions)
}
\arguments{
\item{df}{a dataframe with name,days,products, replicates as first columns. Further columns corresponds to the intensity relatively to ions (one ion by column)}

\item{ions}{vector containing the names of the column to be considered as ions and to be analysed}
}
\description{
returns dataframe containing the average (avg) by product and day
}
