% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvQCGetDfs.r
\name{ptrvQCGetDfs}
\alias{ptrvQCGetDfs}
\title{ptrvQCGetDfs}
\usage{
ptrvQCGetDfs(repo, ncharDays = c(9, 10), ncharProd = c(12, 15), ncharRep = 17)
}
\arguments{
\item{repo}{a repository (with / instead of \) where all the .txt are saved}

\item{ncharDays}{a vector of two numbers representing the positions of first (and last) characters corresponding to the day in the file name}

\item{ncharProd}{a vector of two numbers representing the positions of first (and last) characters  corresponding to the product in the file name}

\item{ncharRep}{one number representing the position of the character corresponding to the replicate in the file name}
}
\value{
a list of 3 data frames containing: area under curves (auc), maximal intensity reached (max), time when the maximal intensity is reached (tmax)
}
\description{
returns and saves of 3 data frames containing: area under curves (auc), maximal intensity reached (max), time when the maximal intensity is reached (tmax)
}
