% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvQCKruskal.r
\name{ptrvQCKruskal}
\alias{ptrvQCKruskal}
\title{ptrvQCKruskal}
\usage{
ptrvQCKruskal(df)
}
\arguments{
\item{df}{a dataframe with name,days,products, replicates as first columns. Further columns corresponds to the intensity relatively to ions (one ion by column)}
}
\value{
results of kruskal wallis test on days
}
\description{
ptrvQCKruskal
}
