% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvQCPlotSummary.r
\name{ptrvQCPlotSummary}
\alias{ptrvQCPlotSummary}
\title{ptrvQCPlotSummary}
\usage{
ptrvQCPlotSummary(df, ions = NULL)
}
\arguments{
\item{df}{a dataframe with name,days,products, replicates as first columns. Further columns corresponds to the intensity relatively to ions (one ion by column)}

\item{ions}{vector containing the names of the column to be considered as ions and to be analysed}
}
\value{
a list of plots (bars, boxplot, points)
}
\description{
ptrvQCPlotSummary
}
