% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvQCStat.r
\name{ptrvQCStats}
\alias{ptrvQCStats}
\title{ptrvQCStats}
\usage{
ptrvQCStats(df, ions = NULL)
}
\arguments{
\item{df}{a dataframe with name,days,products, replicates as first columns. Further columns corresponds to the intensity relatively to ions (one ion by column)}

\item{ions}{vector containing the names of the column to be considered as ions and to be analysed}
}
\value{
a list of dataframes: avg for the average, std for standard deviations and cv for the coefficient of variations
}
\description{
returns and saves a list of dataframes containing the average (avg), standard deviations (std) and coefficients of variation (cv)
}
