% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvSameTimePoints.r
\name{ptrvSameTimePoints}
\alias{ptrvSameTimePoints}
\title{ptrvSameTimePoints
returns a dataframe with the same timepoints}
\usage{
ptrvSameTimePoints(
  df,
  npoints = 100,
  meanBySubject = TRUE,
  meanByRep = FALSE,
  normalizeByTime = FALSE,
  breakTimes = NULL
)
}
\arguments{
\item{df}{a data frame whose colnames are time, product, subject,rep, duration and intensity}

\item{npoints}{number of timepoints to be obtained}

\item{meanBySubject}{boolean indicating whether the mean by subject should be computed}

\item{meanByRep}{boolean indicating whether the mean by replicate should be computed}

\item{normalizeByTime}{boolean indicating whether the data should be normalized (T) or not}

\item{breakTimes}{indicating the breaking time points. Default to seq(0,max(df$time),length.out=npoints)}
}
\description{
ptrvSameTimePoints
returns a dataframe with the same timepoints
}
