% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBy.R
\name{runBy}
\alias{runBy}
\title{Run analyses by batch.}
\usage{
runBy(
  df,
  runBy = "",
  selection = "",
  outputFormat = "",
  outputFile = "",
  title = "",
  fun,
  parameters = list()
)
}
\arguments{
\item{df}{Dataframe.}

\item{runBy}{Split x in several datasets corresponding to the combinations of each factor of runBy, then run a separate analysis for each dataset.}

\item{selection}{Run the analysis for a subset of x. Use syntax of sqldf.}

\item{outputFormat}{Format of output, "" (no output), "html", "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp, "svg" or "wmf" (windows only)}

\item{outputFile}{Name of the file (if outputFormat !="").}

\item{title}{Title of the output.}

\item{fun}{Function}

\item{parameters}{Parameters for fun.}
}
\value{
List.
}
\description{
Run analyses by batch.
}
