% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameTimePoints.r
\name{sameTimePoints}
\alias{sameTimePoints}
\title{sameTimePoints
sameTimePoints}
\usage{
sameTimePoints(
  df,
  npoints = 100,
  colTime = "time",
  colDuration = "duration",
  colIntensity = "intensity",
  breakTimes = NULL
)
}
\arguments{
\item{df}{data frame with at least three columns (named as colTime, colDuration and colIntensity)}

\item{npoints}{number of points (to fill breakTimes automatically)}

\item{colTime}{name of the column corresponding to time}

\item{colDuration}{name of the column corresponding to duration}

\item{colIntensity}{name of the column corresponding to intensity}

\item{breakTimes}{vector of numerics corresponding to time breaks}
}
\description{
sameTimePoints
sameTimePoints
}
