% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothRes.r
\name{smoothRes}
\alias{smoothRes}
\title{smoothRes
Returns a list of all graph smoothed}
\usage{
smoothRes(df, smooth = 0.7)
}
\arguments{
\item{df}{data.frame with ion, time, intensity as columns}

\item{smooth}{number between 0 and 1 indicating the degree of smoothing -same as smooth.spline}
}
\description{
smoothRes
Returns a list of all graph smoothed
}
