% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.listPtrv.r
\name{totalCVA}
\alias{totalCVA}
\title{cva.listPtrv}
\usage{
totalCVA(
  df,
  ionToRemove = NULL,
  option = "OneWayANOVA",
  normalizeByEval = FALSE,
  log = FALSE,
  axes = c(1, 2)
)
}
\arguments{
\item{df}{data frame with "ion","intensity","product","subject" columns,}

\item{ionToRemove}{vector of character containing ions to remove from the analysis}

\item{normalizeByEval}{Boolean. By default TRUE}

\item{log}{boolean indicated if the intensities should be logged (TRUE) or not (FALSE)}

\item{type}{Choose among: "CVA of abundances","ANOVA of intensities","PCA of abundances","Intensity Heatmap"}
}
\description{
cva.listPtrv
}
