% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.listPtrv.r
\name{totalPCA}
\alias{totalPCA}
\title{analysis.listPtrv}
\usage{
totalPCA(
  df,
  ionToRemove = NULL,
  dataType = "productMeans",
  normalizeByEval = FALSE,
  log = FALSE,
  axes = list(c(1, 2)),
  representation = "TwoMaps",
  option = "Correlation",
  variableName = "ion",
  scoreName = "intensity",
  expandBiplot = NULL
)
}
\arguments{
\item{df}{data frame with "ion","intensity","product","subject" columns,}

\item{ionToRemove}{vector of character containing ions to remove from the analysis}

\item{normalizeByEval}{Boolean. By default TRUE}

\item{log}{boolean indicated if the intensities should be logged (TRUE) or not (FALSE)}

\item{type}{Choose among: "CVA of abundances","ANOVA of intensities","PCA of abundances","Intensity Heatmap"}
}
\description{
analysis.listPtrv
}
