% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCycles.r
\name{detectCycle}
\alias{detectCycle}
\title{detectCycle
Detects breathing cycles (negative peaks with smoothing methods)}
\usage{
detectCycle(
  df,
  maxPeaks = NULL,
  smoothMethod = "MovingAverage",
  method = "MAD",
  halfWindowSize = 5,
  maximum = NULL,
  SNR = 0,
  minimalDuration = 2
)
}
\arguments{
\item{df}{dataframe with colnames as 'intensity','time'}

\item{maxPeaks}{minimal intensity of the original curve (not the opposite) required to be considered as a peak}

\item{smoothMethod}{smoothing method. Among "MovingAverage" or "SavitzkyGolay". See MSnbase::smooth for more details.}

\item{method}{method to estimate the noise "MAD" or ""}

\item{halfWindowSize}{required for the smoothing method. Integer corresponding to the half window size to consider.}

\item{maximum}{last value to attribue to the breathing cycle (end of tasting for example)}

\item{SNR}{signal noise ratio (0 by default) required to be considered as a peak}

\item{minimalDuration}{minimalDuration of a breathing cycle (2 by default - as a accelerated breathing frequency corresponds to 20 cycles and more by minut and a low frequency is 12 and less -)}
}
\description{
detectCycle
Detects breathing cycles (negative peaks with smoothing methods)
}
