% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByEval.r
\name{normalizeByEval}
\alias{normalizeByEval}
\title{normalizeByEval
Divides each evaluation by the sum of its variables}
\usage{
normalizeByEval(df, ionToUse = NULL)
}
\arguments{
\item{df}{data frame containing the column "file","product","subject","rep","ion","intensity" (optionally "file" and "time")}

\item{ionToUse}{if NULL, all the ions are use to normalize, if a vector of names of ions, it is used to normalize}
}
\description{
normalizeByEval
Divides each evaluation by the sum of its variables
}
