% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvIntensity.r
\name{ptrvIntensity}
\alias{ptrvIntensity}
\title{ptrvIntensity}
\usage{
ptrvIntensity(result, timePeriod = NULL, negativeAsNull = TRUE)
}
\arguments{
\item{result}{is a result of ptrvIntensityByTime$res}

\item{timePeriod}{vector with two numbers indicating an interval to calculate the intensities}

\item{negativeAsNull}{If TRUE, all negative value are replaced by 0}
}
\description{
ptrvIntensity
}
