% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvComparableData.r
\name{ptrvListComparableData}
\alias{ptrvListComparableData}
\title{ptrvListComparableData}
\usage{
ptrvListComparableData(df, metaData, timeColumn = "time")
}
\arguments{
\item{df}{data.frame with time column}

\item{metaData}{metaData containing these columns: "finPTR_s","miseEnBouche_s"}

\item{timeColumn}{name of the time column in df}
}
\description{
ptrvListComparableData
}
