% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvListIntensity.r
\name{ptrvListIntensity}
\alias{ptrvListIntensity}
\title{ptrvListIntensity}
\usage{
ptrvListIntensity(
  df,
  format = "wide",
  stat = "area",
  timePeriod = NULL,
  negativeAsNull = TRUE,
  fill = NULL
)
}
\arguments{
\item{df}{a result of ptrvListIntensityByFiles ($res) ie a data.frame whose colnames are  'time ' 'ion' 'intensity' 'duration' 'file' and others}

\item{format}{"wide" by default, can also be 'long'}

\item{stat}{"area" by default but can also be "tmax" or "max"}

\item{timePeriod}{calculates on a specific time period. If not null, can be a vector with two values (start and stop) or a data.frame whose colnames are file, start and stop}

\item{negativeAsNull}{If TRUE, all negative value are replaced by 0}
}
\description{
ptrvListIntensity
}
