% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvSignificantSNRIons.r
\name{ptrvSignificantSNRIons}
\alias{ptrvSignificantSNRIons}
\title{ptrvSignificantSNRIons}
\usage{
ptrvSignificantSNRIons(
  dataset,
  method = "max",
  referenceBreath,
  noisePeriod = c(0, 30),
  correction = "cycle",
  multiplyNoiseBy = 3,
  removeNoise = FALSE,
  minimalDuration = 2,
  maxPeaks = NULL
)
}
\arguments{
\item{dataset}{whose names are IN THIS ORDER:AbsTime, RelTime, Cycle, Ion1,... ,Ionp.}

\item{method}{"max" or "ttest" (TODO "tangente")}

\item{referenceBreath}{name of the ion taken as reference for breathing}

\item{noisePeriod}{vector with two numbers indicating the period corresponding to a blank period.}

\item{correction}{"none" or "cycle". See Details.}

\item{multiplyNoiseBy}{a number such as a ion whose maximal intensity during tasting/maximal intensity during blank period > multiplyNoiseBy is detected as significant (used for method=="max")}

\item{removeNoise}{if TRUE, the noise is removed in the calculation of intensity by time}

\item{minimalDuration}{minimalDuration of a breathing cycle (2 by default - as a accelerated breathing frequency corresponds to 20 cycles and more by minut and a low frequency is 12 and less -)}

\item{maxPeaks}{A number. Below this value, a peak is not considered as a peak}
}
\description{
ptrvSignificantSNRIons
}
