% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrvSignificantSNRIons.r
\name{ptrvSignificantSNRIons}
\alias{ptrvSignificantSNRIons}
\title{ptrvSignificantSNRIons}
\usage{
ptrvSignificantSNRIons(
  dataset,
  method = "max",
  referenceBreath,
  noisePeriod = c(0, 30),
  correction = "cycle",
  multiplyNoiseBy = 3,
  removeNoise = FALSE,
  minimalDuration = 2,
  halfWindowSize = 5,
  smoothMethod = "MovingAverage",
  minExpi = NULL,
  maxInspi = NULL,
  forMinExpiDivideMaxIntBy = 5,
  forMaxInspiDivideMaxIntBy = 4
)
}
\arguments{
\item{dataset}{whose names are timeCol, Ion1,... ,Ionp.}

\item{method}{"max" or "ttest" (TODO "tangente")}

\item{referenceBreath}{name of the ion taken as reference for breathing}

\item{noisePeriod}{vector with two numbers indicating the period corresponding to a blank period.}

\item{correction}{"none" or "cycle". See Details.}

\item{multiplyNoiseBy}{a number such as a ion whose maximal intensity during tasting/maximal intensity during blank period > multiplyNoiseBy is detected as significant (used for method=="max")}

\item{removeNoise}{if TRUE, the noise is removed in the calculation of intensity by time}

\item{minimalDuration}{minimalDuration of a breathing cycle (2 by default - as a accelerated breathing frequency corresponds to 20 cycles and more by minut and a low frequency is 12 and less -)}

\item{halfWindowSize}{parameter used to determine the smooth for the peak-picking used in break detection}

\item{smoothMethod}{NULL, "SavitzkyGolay" or "MovingAverage"}

\item{minExpi}{minimal value for an expiry intensity to be detected as an expiry (blue dotted line in the plot).}

\item{maxInspi}{maximal intensity of the original curve required to be considered as a breathing peak (purple dotted line in the plot)}

\item{forMinExpiDivideMaxIntBy}{default to 5. When minExpi is null, it is evaluated as the ratio of the maximal intensity and forMinIntensityDivideMaxIntBy}

\item{forMaxInspiDivideMaxIntBy}{default to 5. When minExpi is null, it is evaluated as the ratio of the maximal intensity and forMinIntensityDivideMaxIntBy}
}
\description{
ptrvSignificantSNRIons
}
